Remove `fetch_disp`, `_netcdf`, `_reader` modules + tests
