+++ src/opera_utils/disp/__init__.py
@@ -2,26 +2,16 @@
 
 from __future__ import annotations
 
-from . import _reader as reader
 from ._product import DispProduct, DispProductStack
-from ._rebase import rebase_timeseries
+from ._remote import open_file, open_h5
 from ._search import search
-
-# Remote access is based on optional dependencies
-try:
-    from ._netcdf import save_data
-    from ._remote import open_file, open_h5
-except ImportError:
-    pass
-
+from ._xarray import create_rebased_stack
 
 __all__ = [
     "DispProduct",
     "DispProductStack",
+    "create_rebased_stack",
     "open_file",
     "open_h5",
-    "reader",
-    "rebase_timeseries",
-    "save_data",
     "search",
 ]
